move `/users` endpoint to `/user/list`
