+++ api/src/paths/user/list.test.ts
@@ -2,10 +2,10 @@ import chai, { expect } from 'chai';
 import { describe } from 'mocha';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
-import * as db from '../database/db';
-import { UserService } from '../services/user-service';
-import { getMockDBConnection, getRequestHandlerMocks } from '../__mocks__/db';
-import * as users from './users';
+import * as db from '../../database/db';
+import { UserService } from '../../services/user-service';
+import { getMockDBConnection, getRequestHandlerMocks } from '../../__mocks__/db';
+import * as users from './list';
 
 chai.use(sinonChai);
